Mostramos los productos del carrito y los datos del usuario en el resumen del pedido

+++ src/main/java/edu/unam/integrador/services/IUsuarioService.java
@@ -6,7 +6,7 @@ import edu.unam.integrador.model.Usuario;
 
 public interface IUsuarioService {
     List<Usuario> findAll();
-    Optional<Usuario> findById(Integer id);
+    Optional<Usuario> findById(Long id);
     Usuario save (Usuario usuario);
     Optional<Usuario> findByEmail(String email);
 }
